feat: reimplement legacy symbol parsing system
