+++ pyproject.toml
@@ -2,6 +2,7 @@
 requires = [
     "scikit-build-core >= 0.4.3",
     "nanobind >= 1.3.2",
+    "typing_extensions",
 ]
 build-backend = "scikit_build_core.build"
